Ensure that directory paths passed from Bookmarks through CommandLauncher to Windows Explorer are properly quoted
- Quote parameters passed to Windows Explorer.
- Extract a new ShellArgumentBuilder class, shared by all extensions, to build and quote command-line arguments consistently.
